# FoodApp/Documentation/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FoodApp API",
    "version": "v1"
  },
  "paths": {
...
    "/api/meals": {
      "get": {
        "tags": [
          "Meal"
        ],
        "summary": "Retrieves a list of meals with optional filtering and pagination.",
        "description": "Requires user authentication. Supports pagination with page and perPage parameters.\r\nIf name is provided, meals containing that text in their name are returned.\r\nResults are sorted alphabetically by name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Optional name to search for (partial match, case-insensitive).",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (default: 1).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "Number of items per page (default: 25, max: 100).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the list of matching meals."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          }
        }
      },
      "post": {
        "tags": [
          "Meal"
        ],
        "summary": "Creates a new meal record.",
        "description": "Requires admin role authentication. The meal name must be unique within the system.\r\nMeal names are trimmed of whitespace and validated for minimum length.\r\nCommon meal types include Breakfast, Lunch, Dinner, Snack, but custom types are supported.",
        "requestBody": {
          "description": "The meal data to create. Name is required and must be unique.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the created meal."
          },
          "400": {
            "description": "Bad Request - invalid meal data or missing required fields."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      }
    },
    "/api/meals/{id}": {
      "get": {
        "tags": [
          "Meal"
        ],
        "summary": "Retrieves a specific meal by its ID.",
        "description": "Requires user authentication. Returns detailed information about a single meal.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the meal.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested meal."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "404": {
            "description": "Not Found - meal with the specified ID does not exist."
          }
        }
      },
      "patch": {
        "tags": [
          "Meal"
        ],
        "summary": "Updates an existing meal record.",
        "description": "Requires admin role authentication. Only the name field can be updated.\r\nThe updated name must be unique within the system.\r\nReturns 400 if no fields are provided for update.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the meal to update.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated meal."
          },
          "400": {
            "description": "Bad Request - no valid fields provided for update or meal not found."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      },
      "delete": {
        "tags": [
          "Meal"
        ],
        "summary": "Deletes a meal by its ID.",
        "description": "Requires admin role authentication. This operation performs a cascading update:\r\n- The meal is permanently removed from the database\r\n- All recipes that had this meal association will have their meal references removed\r\n- This is a destructive operation that cannot be undone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the meal to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns deletion status (true if successful, false if not found)."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MealCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put **_ONLY_** your JWT Bearer token on textbox below!",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}